requirements.txt
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0
huggingface-hub>=0.19.0
safetensors>=0.4.0
tqdm>=4.66.0
pyyaml>=6.0
einops>=0.7.0
